python-agents/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.26.0
openai==1.10.0
langchain==0.1.0
pandas==2.1.4
numpy==1.26.0
